@@ -9,7 +9,8 @@
     "test": "npm run-script lint && ./bin/test.sh && ./bin/test-ui.sh --single-run",
     "prepublish": "npm test && ./bin/build.sh",
     "start": "./bin/devserver.sh",
-    "build": "NODE_ENV=production ./node_modules/webpack/bin/webpack.js -p --display-chunks"
+    "build": "NODE_ENV=production ./node_modules/webpack/bin/webpack.js -p --display-chunks",
+    "docs": "./cli/index.js --config=doc/tinydoc.conf.js"
   },
   "repository": {
     "type": "git",
@@ -35,6 +36,7 @@
     "fs-extra": "0.18.0",
     "git-log-parser": "^1.1.0",
     "glob": "5.0.3",
+    "highlight.js": "^8.8.0",
     "less": "2.3.1",
     "less-loader": "2.0.0",
     "lodash": "3.6.0",
@@ -44,6 +46,7 @@
     "react": "^0.13.3",
     "style-loader": "0.8.3",
     "url-loader": "0.5.5",
+    "weakmap": "0.0.6",
     "weakset": "^0.2.1",
     "webpack": "1.7.3"
   },
@@ -51,8 +54,8 @@
     "babel-core": "5.3.3",
     "babel-loader": "5.0.0",
     "chai": "2.1.0",
-    "eslint": "^1.3.1",
-    "eslint-plugin-react": "^3.2.2",
+    "eslint": "^1.4.3",
+    "eslint-plugin-react": "^3.4.1",
     "highlight.js": "^8.4.0",
     "imports-loader": "^0.6.4",
     "jquery": "^2.1.3",
@@ -67,6 +70,7 @@
     "react": "0.13.3",
     "react-hot-loader": "",
     "react-router": "",
+    "recast": "^0.10.32",
     "webpack-dev-middleware": "1.0.11",
     "webpack-dev-server": "1.7.0"
   }
