@@ -17,6 +17,7 @@ const markedOptions = Object.freeze({
 });
 
 hljs.registerLanguage('javascript', require('highlight.js/lib/languages/javascript'));
+hljs.registerLanguage('xml', require('highlight.js/lib/languages/xml'));
 hljs.registerLanguage('bash', require('highlight.js/lib/languages/bash'));
 hljs.registerLanguage('ruby', require('highlight.js/lib/languages/ruby'));
 hljs.registerLanguage('css', require('highlight.js/lib/languages/css'));
@@ -33,7 +34,10 @@ var MarkdownText = React.createClass({
   },
 
   propTypes: {
-    linkify: React.PropTypes.bool
+    linkify: React.PropTypes.bool,
+    jumpy: React.PropTypes.bool,
+    children: React.PropTypes.any,
+    className: React.PropTypes.string,
   },
 
   getDefaultProps: function() {
@@ -69,6 +73,16 @@ var MarkdownText = React.createClass({
     if (nodes.length > 0) {
       for (var i = 0; i < nodes.length; i=i+1) {
         hljs.highlightBlock(nodes[i]);
+
+        // a hack for better JSX support
+        if (nodes[i].className.match(/xml|javascript/)) {
+          // nodes[i].className = 'javascript';
+          // hljs.highlightBlock(nodes[i]);
+          nodes[i].className = 'xml';
+          hljs.highlightBlock(nodes[i]);
+
+          nodes[i].className = 'hljs xml javascript';
+        }
       }
     }
   },
