@@ -1,43 +1,46 @@
-var VERBOSE = !!process.env.VERBOSE;
-var DEBUG = !!process.env.DEBUG;
+var LEVEL = 'log';
 
 function Logger(context) {
   var logger = {};
   var loggingContext = context + ':';
 
-  var logLevels = [ 'log', 'info', 'warn', 'error' ];
+  var logLevels = [ 'debug', 'info', 'log', 'warn', 'error' ];
 
   logLevels.forEach(function(logLevel) {
     logger[logLevel] = function() {
-      if (!VERBOSE && logLevels.indexOf(logLevel) < 2) {
+      var args;
+
+      if (logLevels.indexOf(logLevel) < logLevels.indexOf(LEVEL)) {
         return;
       }
 
-      var args = [].slice.call(arguments);
+      args = [].slice.call(arguments);
       args[0] = '[' + logLevel + '] ' + loggingContext + ' ' + args[0];
-      console[logLevel].apply(console, args);
+
+      (console[logLevel] || console.log).apply(console, args);
     };
   });
 
-  logger.debug = function() {
-    if (DEBUG) {
-      var args = [].slice.call(arguments);
-      args[0] = '[DEBUG] ' + loggingContext + ' ' + args[0];
-      console.log.apply(console, args);
-    }
-  };
-
   logger.assert = console.assert.bind(console);
+  logger.raw = console;
 
   return logger;
 }
 
-Logger.setVerbose = function(onOrOff) {
-  VERBOSE = onOrOff;
+Logger.setVerbose = function() {
+  LEVEL = 'info';
 };
 
-Logger.setDebug = function(onOrOff) {
-  DEBUG = onOrOff;
+Logger.setDebug = function() {
+  LEVEL = 'debug';
 };
+
+Logger.setLevel = function(level) {
+  LEVEL = level;
+};
+
+if (process.env.DEBUG === '1') {
+  Logger.setDebug();
+}
 
 module.exports = Logger;
