@@ -1,45 +1,39 @@
 var findWhere = require('lodash').findWhere;
 var Logger = require('../../lib/Logger');
-var console = new Logger('cjs');
+var console = new Logger('cjs-linker');
 
 module.exports = function(docs) {
   var indices = [];
 
   docs.forEach(function(doc) {
-    var context;
+    var index;
     var paths = [];
 
     if (doc.isModule) {
-      context = { id: doc.id };
+      index = { id: doc.id };
 
       paths.push(doc.id);
       paths.push(doc.name);
     }
-    else if (doc.ctx) {
-      switch(doc.ctx.type) {
-        case 'method':
-        case 'function':
-        case 'property':
-          var parentDoc = findWhere(docs, { id: doc.ctx.receiver });
+    else {
+      var parentDoc = findWhere(docs, { id: doc.receiver });
 
-          if (!parentDoc) {
-            console.warn('Unable to find parent "%s" for entity "%s"', doc.ctx.receiver, doc.id);
-          }
-          else {
-            context = {
-              id: doc.id,
-              parent: parentDoc.id
-            };
+      if (!parentDoc) {
+        console.warn('Unable to find parent "%s" for entity "%s"', doc.receiver, doc.id);
+      }
+      else {
+        index = {
+          id: doc.id,
+          parent: parentDoc.id
+        };
 
-            paths.push([ parentDoc.id, doc.name ].join(doc.symbol));
-            paths.push([ parentDoc.name, doc.name ].join(doc.symbol));
-          }
-          break;
+        paths.push([ parentDoc.id, doc.name ].join(doc.ctx.symbol));
+        paths.push([ parentDoc.name, doc.name ].join(doc.ctx.symbol));
       }
     }
 
-    if (context) {
-      context.type = 'cjs';
+    if (index) {
+      index.type = 'cjs';
 
       if (doc.alias) {
         paths.push(doc.alias);
@@ -47,7 +41,7 @@ module.exports = function(docs) {
       }
 
       paths.forEach(function(path) {
-        indices.push({ path: path, context: context });
+        indices.push({ path: path, index: index });
       });
     }
   });
