@@ -3,10 +3,20 @@ var extend = require('lodash').extend;
 module.exports = function(database, config, utils, done) {
   var runtimeConfig = extend({}, config, { database: database });
 
-  utils.writeAsset(
-    'plugins/cjs-config.js',
-    'window["cjs-config"]=' + JSON.stringify(runtimeConfig) + ';'
-  );
+  try {
+    utils.writeAsset(
+      'plugins/cjs-config.js',
+      'window["cjs-config"]=' + JSON.stringify(runtimeConfig) + ';'
+    );
 
-  done();
+    done();
+  }
+  catch(e) {
+    if (e.name === 'TypeError' && e.message.match(/circular/)) {
+      console.error(runtimeConfig);
+    }
+
+    throw e;
+  }
+
 };
