@@ -3,39 +3,65 @@ var expect = require('chai').expect;
 
 describe('markdown/scan/parseSections', function() {
   it('should work', function() {
-    var fixture = TestUtils.loadFixture('markdown/fixture.md');
+    var fixture = TestUtils.getInlineString(function() {
+      // # Testing CommonJS Modules
+      //
+      // ## Second-level
+      //
+      // ### Third-level
+      //
+      // #### Fourth-level header
+    }, true);
+
     var sections = parseSections(fixture);
 
-    expect(sections.length).to.equal(5);
+    expect(sections.length).to.equal(2);
     expect(sections).to.contain({
       level: 2,
-      title: 'A little about Mocha'
+      title: 'Second-level'
     });
 
     expect(sections).to.contain({
       level: 3,
-      title: 'Test helpers'
+      title: 'Third-level'
+    });
+
+    expect(sections).not.to.contain({
+      level: 4,
+      title: 'Fourth-level'
     });
   });
 
   it('should work with stroked headings', function() {
-    var fixture = TestUtils.loadFixture('markdown/fixture_with_stroked_headings.md');
+    var fixture = TestUtils.getInlineString(function() {
+      // Testing CommonJS Modules
+      // ========================
+      //
+      // Second-level
+      // ------------
+      //
+      // ### Third-level
+      //
+      // #### Fourth-level header
+    }, true);
+
     var sections = parseSections(fixture);
 
-    expect(sections.length).to.equal(4);
-    expect(sections).to.contain({
-      level: 2,
-      title: 'What you do'
-    });
+    expect(sections.length).to.equal(2);
 
     expect(sections).to.contain({
       level: 2,
-      title: 'How you do it'
+      title: 'Second-level'
     });
 
     expect(sections).to.contain({
       level: 3,
-      title: 'Poking around in the console'
+      title: 'Third-level'
+    });
+
+    expect(sections).not.to.contain({
+      level: 4,
+      title: 'Fourth-level'
     });
   });
 });
