@@ -3,12 +3,28 @@ var expect = require('chai').expect;
 
 describe('markdown/scan/parseTitle', function() {
   it('should work', function() {
-    var fixture = TestUtils.loadFixture('markdown/fixture.md');
+    var fixture = TestUtils.getInlineString(function() {
+      // # Testing CommonJS Modules
+      //
+      // Something.
+      //
+      // ## A little about Mocha
+    }, true);
+
     expect(parseTitle(fixture)).to.equal('Testing CommonJS Modules');
   });
 
   it('should work with stroked headings', function() {
-    var fixture = TestUtils.loadFixture('markdown/fixture_with_stroked_headings.md');
+    var fixture = TestUtils.getInlineString(function() {
+      // Support Week
+      // ============
+      //
+      // What you do
+      // -----------
+      //
+      // Monitor #donk_users, work on high priority bugs and help with reviews.
+    }, true);
+
     expect(parseTitle(fixture)).to.equal('Support Week');
   });
 });
